Moved the bdist_wheel module back to its original location
