+++ setup.cfg
@@ -50,7 +50,7 @@ types =
 console_scripts =
     wheel = wheel._cli:main
 distutils.commands =
-    bdist_wheel = wheel._bdist_wheel:bdist_wheel
+    bdist_wheel = wheel.bdist_wheel:bdist_wheel
 
 [tool:isort]
 src_paths = src
