+++ tests/test_macosx_libfile.py
@@ -8,8 +8,8 @@ from typing import Any, TypeVar
 
 from pytest import CaptureFixture, MonkeyPatch
 
-from wheel._bdist_wheel import get_platform
 from wheel._macosx_libfile import extract_macosx_min_system_version
+from wheel.bdist_wheel import get_platform
 
 T = TypeVar("T")
